+++ mooda/__init__.py
@@ -1,4 +1,8 @@
 from mooda.waterframe import WaterFrame
-from mooda.plotmap import PlotMap
+try:
+    from mooda.plotmap import PlotMap
+except ImportError:
+    print("Warning: You do not have installed the library basemap. You cannot",
+          "use PlotMap.")
 
 name = "mooda"