+++ mooda/waterframe.py
@@ -249,13 +249,20 @@ class WaterFrame:
         else:
             df = df.resample(average_time).mean()
 
-        ax = df[key].plot.bar(ax=ax, legend=True)
-
-        # Write axes
-        ax.set_ylabel(self.meaning[key]['units'])
-        if average_time == 'A':
-            ax.set_xticklabels([format_year(x.get_text())
-                                for x in ax.get_xticklabels()])
+        if isinstance(key, list):
+            ax = df[key].plot.bar(ax=ax, legend=True)
+        else:
+            ax = df[key].plot.bar(ax=ax)
+             # Write axes
+            try:
+                ax.set_ylabel(self.meaning[key]['units'])
+            except KeyError:
+                print("Warning: We don't know the units of", key,
+                      "Please, add info into self.meaning[", key,"['units']")
+            
+            if average_time == 'A':
+                ax.set_xticklabels([format_year(x.get_text())
+                                    for x in ax.get_xticklabels()])
 
         return ax