+++ docs/news.md
@@ -11,6 +11,14 @@ Given a version number MAJOR.MINOR.PATCH, increment the:
 
 These are new features and improvements of note in each release.
 
+## mooda v0.1.0 (Work in progress)
+
+* In mooda/\_\_init\_\_.py: It does not load plotmap.py if you do not have installed the basemap library. We added a warning message.
+* In mooda/waterframe.py:
+  * barplot(): It creates the graph even without knowing the meaning of the parameters.
+* In mooda/access/egim.py:
+  * to_csv(): New function. It creates a CSV file following the OceanSites standard.
+
 ## oceanobs is now mooda - mooda v0.0.1 (July, 2018)
 
 Due to the requirements of the EMSODEV project, the package had to be renamed to "mooda." So, the directory where the MOODA app was, is now called as "mooda_gui."  The release count will start again from scratch, calling the first release as "mooda v0.0.1."